PROTRACTOR TIME OUT FIX
Working with Alumni helpers, fixed issue.
The expect was not running since it was waiting for angular to sync.
For a currently unknown reason angular is not synchronizing.
We have added a 'browser.ignorSynchronization()', to our function and a wait, the test now passes.
Feels pretty hacked together but it means we can move forwards.
Post that really helped can be found at:
	https://github.com/lynndylanhurley/ng-token-auth/issues/182

Currently registering a user is in as a function in the signing in spec.
Need a spec to test registering a user, then refactor the sign in spec to put registering a user in a 'beforeEach'.